add authUser to prevent not admin user form adding or editing or deleting product

+++ src/app/api/admin/delete-product/route.js
@@ -1,4 +1,5 @@
 import connectToDb from "@/database";
+import { AuthUser } from "@/authUser/AuthUser";
 import Product from "@/models/product";
 import { NextResponse } from "next/server";
 
@@ -7,26 +8,34 @@ export const dynamic = "force-dynamic";
 connectToDb();
 export async function DELETE(req) {
   try {
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
-    if (!id) {
-      return NextResponse.json({
-        success: false,
-        message: "product id is required",
-      });
-    } else {
-      const deletedProduct = await Product.findOneAndDelete(id);
-      if (deletedProduct) {
-        return NextResponse.json({
-          success: true,
-          message: "product is deleted successfully",
-        });
-      } else {
+    const authenticatedUser = await AuthUser(req);
+    if (authenticatedUser?.role === "admin") {
+      const { searchParams } = new URL(req.url);
+      const id = searchParams.get("id");
+      if (!id) {
         return NextResponse.json({
           success: false,
-          message: "failed to delete product",
+          message: "product id is required",
         });
+      } else {
+        const deletedProduct = await Product.findOneAndDelete(id);
+        if (deletedProduct) {
+          return NextResponse.json({
+            success: true,
+            message: "product is deleted successfully",
+          });
+        } else {
+          return NextResponse.json({
+            success: false,
+            message: "failed to delete product",
+          });
+        }
       }
+    } else {
+      return NextResponse.json({
+        success: false,
+        message: "sorry you are not allowed to delete this product",
+      });
     }
   } catch (error) {
     return NextResponse.json({
